requirements.txt
Django==3.0.8
whitenoise==5.1.0
python-dotenv==0.14.0
django-crispy-forms==1.9.2
gunicorn==20.0.4
django-braces==1.14.0
